requirements.txt
python-dotenv
groq
pdfplumber